pyproject.toml
[tool.poetry]
name = "finance-assistant-backend"
version = "0.1.0"
description = ""
authors = ["karachunYaraslau <karacuna396@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {version="^0.116.2", extras=["standard"]}
pydantic-settings = "^2.4.0"
sqlalchemy = "^2.0"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
httpx = "^0.27.0"
orjson = "^3.10.7"
loguru = "^0.7.2"
psycopg = {extras = ["binary"], version = "^3.2.10"}
authx = "^1.4.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.1"
bcrypt = "4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"
ruff = "^0.6.9"
black = "^24.8.0"
mypy = "^1.11.0"
types-python-dateutil = "*"
pre-commit = "^3.8.0"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
lint.select = ["E","F","I","UP","B","SIM"]
fix = true

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
